backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
# Motor 3.6+ 才兼容新版 PyMongo（4.9+），否则会出现 _QUERY_OPTIONS 导入错误
motor>=3.6.0,<4.0.0
redis[hiredis]==5.0.8
apscheduler==3.10.4
httpx==0.27.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
